Uploaded project to GitHub
I decided to upload this little vanilla js test I did a couple of days ago to the community, so you can see how easy it is to create a complete note taking app with vanilla js using as little as 100 lines of code.

Hope you enjoy it the same I did when I worked on it!